requirements.txt
streamlit==1.49.0
firebase-admin==7.2.0
google-generativeai==0.8.6
python-dotenv==1.0.1
streamlit-agraph==0.0.45
